requirements.txt
mido==1.3.0
numpy==1.25.2
packaging==23.1
PyYAML==6.0.2
simpleaudio==1.0.4
